normalize comments links
in short articles because those articles aren't cached and the comment
link text contains the current number of contents. Otherwise, such
irrelevant miniscule variation in the number of posted comments would
lead to an updated checksum for the entry which in turn would make it
appear as a new article in a feed-reader.